requirements.txt
yt-dlp
pydub
pygraphviz
transformers
openai-whisper
streamlit
Pillow
openai